pyproject.toml
[tool.poetry]
name = "task-app"
version = "0.1.0"
description = ""
authors = ["Andrei Matei <a.matei2206@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
